requirement.txt
fastapi
uvicorn
langchain
langgraph
openai
pypdf
python-multipart
python-dotenv